+++ fingerprint-auth-demo-backend/src/main/java/com/vunm/demo/domain/service/RequestLogService.java
@@ -85,4 +85,4 @@ public class RequestLogService {
             .stream()
             .toList();
     }
-} 
+}